requirements.txt

pandas
faiss-cpu
pypdf
python-dotenv

langchain
langchain-core
langchain-community
langchain-google-genai


streamlit